app/src/main/java/com/rashidsaleem/notesapp/feature_addNote/domain/DeleteNoteUseCase.kt
package com.rashidsaleem.notesapp.feature_addNote.domain

import com.rashidsaleem.notesapp.core.data.respository.NotesRepositoryImpl
import com.rashidsaleem.notesapp.core.domain.repository.NotesRepository
import kotlinx.coroutines.CoroutineDispatcher
import kotlinx.coroutines.Dispatchers
import kotlinx.coroutines.withContext

class DeleteNoteUseCase(
    private val repository: NotesRepository,
    private val ioDispatcher: CoroutineDispatcher = Dispatchers.IO,
) {


    suspend fun execute(id: Int) {

        withContext(ioDispatcher) {
            repository.delete(id)
        }
    }
}